Implement ia-opt
ia-opt: There is an interval of pixels whose neighborhoods are entirely
contained in an image and therefore no out of bounds checks need to be
done. This is a relatively naive implementation which tries to track
whether the position of the `RectangleNeighborhoodLocalizableSampler` is
in that interval and returns a possibly faster cursor if so.

For cursors, there is a further optimization potential: If you iterate
over the "unsafe" areas first and then over the "safe" area, one would
not need to keep track of whether the position is inside the interval or
not.

Signed-off-by: Squareys <Squareys@googlemail.com>